Merge the is_closed bool into the SearchTreeNode pointer.
This is a bit nasty, but I don't know how to effectively use ~ointers~ and ~ux~
as of today.

- ~ux~ says ~uN~ doesn't really use just `N` bits, so the niche optimisation
might not happen.
- ~ointers~ uses a full-width pointer, but just gives you a tiny buffer inside
of it.

Considering both limitations, it's hard to have a slim pointer NonNull<T> such
that it leaves as many bits free as possible considering it points to a
well-alligned T object.